Report most of staking module
